+++ AGENTS.md
@@ -24,7 +24,7 @@ The `instro` repository is a shared `uv`/`cargo` workspace. The top-level python
 
 | Workspace | Path | What it is |
 |---|---|---|
-| Python/`uv` | `instro/<category>/` | Category code: HAL class (`InstroPSU`, `InstroDMM`, …), `types.py`, the base driver class (`PSUDriverBase`, etc.). Categories: `psu`, `dmm`, `eload`, `scope`, `daq`, `i2c`, `modbus`. |
+| Python/`uv` | `instro/<category>/` | Category code: HAL class (`InstroPSU`, `InstroDMM`, …), `types.py`, the base driver class (`PSUDriverBase`, etc.). Categories: `psu`, `dmm`, `eload`, `scope`, `awg`, `daq`, `i2c`, `modbus`. |
 | Python/`uv` | `instro/<category>/drivers/` | Concrete vendor drivers, one file per vendor/model family. Registered in `drivers/__init__.py`. |
 | Python/`uv` | `instro/lib/transports/` | Transport drivers (`VisaDriver`). Category bases are transport-agnostic; concrete drivers compose transports. |
 | Python/`uv` | `packages/instro-contrib/` | Community-contributed drivers. Mirrors core layout under `instro/contrib/`. |
@@ -114,7 +114,7 @@ This repo prefers duplicated, explicit code over premature abstraction. The cons
 - **`_check_errors` is per-driver because SCPI error semantics vary.** Response prefix (`"0"` for B&K/Rigol, `"+0"` for Siglent), command form (`SYST:ERR?` vs `:SYST:ERR?`), and the vendor name in the raised message are all per-device. A configurable mixin would carry more code than the four duplicated lines.
 - **`_write_checked` is per-driver because some drivers can't use it.** The helper assumes `write + _check_errors` is one atomic step. Stateful drivers can't fit that shape: `BK9140` must hold the VISA lock across `INST <n>` channel-select + write + check, so it inlines the sequence. Keeping the helper driver-local lets stateless drivers stay terse and stateful ones write atomic sequences directly.
 - **`pkgutil.extend_path` in `drivers/__init__.py`** is required for any category whose drivers can come from workspace vendor packages (`daq`, `i2c` currently). Without it, vendor-package subpackages disappear at import time.
-- **Shared JSON-config blocks live in `instro/lib/config.py`.** `TimingConfig`, the publisher config models (`NominalCorePublisherConfig`, `FilePublisherConfig`, `PublisherConfigType`), `build_publisher`, and the generic `load_config(config, model_cls)` loader are identical across category configs, so category `config.py` modules (`instro/psu/config.py`, `instro/dmm/config.py`, `instro/eload/config.py`) import them from there instead of redefining them. Category-specific blocks (the vendor registry, the driver config, DMM's `measurement` block) stay in the category's own `config.py`. Modbus and EtherNet/IP predate this and keep their own divergent `TimingConfig` variants in their own types modules.
+- **Shared JSON-config blocks live in `instro/lib/config.py`.** `TimingConfig`, the publisher config models (`NominalCorePublisherConfig`, `FilePublisherConfig`, `PublisherConfigType`), `build_publisher`, and the generic `load_config(config, model_cls)` loader are identical across category configs, so category `config.py` modules (`instro/psu/config.py`, `instro/dmm/config.py`, `instro/eload/config.py`, `instro/awg/config.py`) import them from there instead of redefining them. Category-specific blocks (the vendor registry, the driver config, DMM's `measurement` block) stay in the category's own `config.py`. Modbus and EtherNet/IP predate this and keep their own divergent `TimingConfig` variants in their own types modules.
 - **VISA drivers' `__init__` accepts `str | VisaConfig`.** `VisaConfig` is the canonical customization vehicle for `VisaDriver`. Don't propose dropping the union. Drivers on other transports take whatever their transport needs.
 - **No vendor-string factory in the direct Python construction API** (`Instrument.create(vendor="bk", ...)`). Construct concrete drivers explicitly and pass them in: `InstroPSU(name="x", driver=BK9115(...), num_channels=1)`. This doesn't extend to JSON/dict config loading (`InstroPSU(config=...)`): resolving a validated `driver.name` string to a class via `PSU_VENDOR_REGISTRY` is an implementation detail of that declarative config format's strict schema, not a public factory method — direct construction stays mandatory and unaffected.
 - **No driver-side facade or back-channel.** Drivers don't hold a reference back to the category HAL. Any vendor-specific state a driver needs across calls (e.g. an `nidaqmx.Task` handle, a `VisaDriver`, a cached sample rate) lives on the driver itself.
