+++ docs/guides/instrumentation/instruments.mdx
@@ -36,7 +36,7 @@ Supported hardware: [National Instruments, LabJack T-Series, Measurement Computi
 Supported hardware: [Keysight 1200X, Tektronix 2-series, and Siglent SDS1000X-E](https://nominal-io.github.io/instro/instruments/scope/).
 
 <Card title="Arbitrary waveform generators" icon="waveform-lines" href="/instrumentation/awg">
-  `InstroAWG`: waveform programming (sine, square, arbitrary, and more) with amplitude, offset, and phase control. **Unstable**: ships in `instro-unstable`.
+  `InstroAWG`: waveform programming (sine, square, arbitrary, and more) with amplitude, offset, and phase control.
 </Card>
 
-Supported hardware: [Rigol DG1022Z (unstable)](https://nominal-io.github.io/instro/instruments/awg/).
+Supported hardware: [Keysight 33521B and Rigol DG1022Z](https://nominal-io.github.io/instro/instruments/awg/).
