+++ docs/guides/instrumentation/overview.mdx
@@ -210,7 +210,7 @@ These are the instrument types available out of the box. The [SDK reference](htt
 - `InstroScope`: oscilloscopes
 - `InstroFlowController`: gas mass-flow controllers
 - `I2CInterface`: I2C bus communication devices
-- `InstroAWG`: arbitrary waveform generators (unstable, install with `instro[unstable]`)
+- `InstroAWG`: arbitrary waveform generators
 - `ModbusDevice`: Modbus register/coil devices (TCP or RTU)
 - `EtherNetIPDevice`: EtherNet/IP and CIP devices, via `instro.ethernetip` (install with `instro[ethernetip]`)
 
@@ -242,8 +242,9 @@ These are the vendor/model-specific drivers available out of the box.
   - Agilent/HP/Keysight 34401A
   - Keysight 34461A
   - Simulated (no hardware required)
-- `InstroAWG` (unstable)
+- `InstroAWG`
   - Rigol DG1022Z
+  - Keysight 33521B
 - `InstroScope`
   - Keysight 1200X
   - Tektronix 2-Series
