+++ docs/reference/src/instruments/awg.md
@@ -1,31 +1,29 @@
 # Arbitrary Waveform Generator (AWG)
 
-This category ships in the `instro-unstable` package (`instro[unstable]`); its API is not settled and may change without notice.
-
 Errors raised by these methods are documented in [Exceptions](../reference/exceptions.md).
 
 ## Interface
 
-::: instro.unstable.awg.InstroAWG
+::: instro.awg.InstroAWG
 
 ## Types & Configuration
 
-::: instro.unstable.awg.types
+::: instro.awg.types
 
 ## Driver Interface
 
-::: instro.unstable.awg.AWGDriverBase
+::: instro.awg.AWGDriverBase
 
 ## Vendor Drivers
 
 ### Keysight 33521B
 
-::: instro.unstable.awg.drivers.keysight_33521b
+::: instro.awg.drivers.keysight_33521b
     options:
       heading_level: 4
 
 ### Rigol DG1022Z
 
-::: instro.unstable.awg.drivers.rigol_dg1022z
+::: instro.awg.drivers.rigol_dg1022z
     options:
       heading_level: 4
