+++ docs/reference/src/instruments/index.md
@@ -12,7 +12,7 @@ instruments. Each instrument type defines a standard API that works across multi
 | [`InstroScope`](scope.md) | Oscilloscopes |
 | [`InstroFlowController`](flowcontroller.md) | Flow Controllers |
 | [`I2CInterface`](i2c.md) | I2C bus communication devices |
-| [`InstroAWG`](awg.md) | Arbitrary waveform generators (unstable) |
+| [`InstroAWG`](awg.md) | Arbitrary waveform generators |
 
 Each instrument page includes the interface, configuration types, driver base classes,
 and vendor-specific driver implementations. Errors raised by instrument methods are
