+++ instro/awg/__init__.py
@@ -1,11 +1,11 @@
 """Arbitrary Waveform Generator (AWG) instrument interface package."""
 
-from instro.unstable.awg.awg import (
+from instro.awg.awg import (
     AWGDriverBase,
     InstroAWG,
 )
-from instro.unstable.awg.config import AWGConfig
-from instro.unstable.awg.types import (
+from instro.awg.config import AWGConfig
+from instro.awg.types import (
     AmplitudeMeasurementUnit,
     Arbitrary,
     BurstTriggerSource,
