+++ instro/awg/drivers/keysight_33521b.py
@@ -5,9 +5,8 @@ from __future__ import annotations
 import logging
 import sys
 
-from instro.lib.transports.visa import VisaConfig, VisaDriver
-from instro.unstable.awg.awg import AWGDriverBase
-from instro.unstable.awg.types import (
+from instro.awg.awg import AWGDriverBase
+from instro.awg.types import (
     AmplitudeMeasurementUnit,
     Arbitrary,
     BurstTriggerSource,
@@ -24,6 +23,7 @@ from instro.unstable.awg.types import (
     Triangle,
     Waveform,
 )
+from instro.lib.transports.visa import VisaConfig, VisaDriver
 
 _HIGH_Z_SENTINEL = 9.9e37
