--- packages/instro-unstable/instro/unstable/awg/drivers/__init__.py
@@ -1,11 +1,0 @@
-"""AWG drivers package."""
-
-from instro.unstable.awg import AWGDriverBase
-from instro.unstable.awg.drivers.keysight_33521b import Keysight33521B
-from instro.unstable.awg.drivers.rigol_dg1022z import RigolDG1022Z
-
-__all__ = [
-    "AWGDriverBase",
-    "Keysight33521B",
-    "RigolDG1022Z",
-]
