+++ tests/awg/keysight/test_keysight_33521b_hardware.py
@@ -7,9 +7,8 @@ from collections.abc import Iterator
 
 import pytest
 
-from instro.lib.transports import VisaConfig
-from instro.unstable.awg.drivers import Keysight33521B
-from instro.unstable.awg.types import (
+from instro.awg.drivers import Keysight33521B
+from instro.awg.types import (
     AmplitudeMeasurementUnit,
     Arbitrary,
     BurstTriggerSource,
@@ -26,6 +25,7 @@ from instro.unstable.awg.types import (
     Triangle,
     Waveform,
 )
+from instro.lib.transports import VisaConfig
 
 pytestmark = pytest.mark.hardware
