+++ tests/awg/keysight/test_keysight_33521b_software.py
@@ -7,9 +7,8 @@ from unittest.mock import MagicMock, call, patch
 
 import pytest
 
-from instro.lib.transports import VisaConfig
-from instro.unstable.awg.drivers import Keysight33521B
-from instro.unstable.awg.types import (
+from instro.awg.drivers import Keysight33521B
+from instro.awg.types import (
     AmplitudeMeasurementUnit,
     Arbitrary,
     BurstTriggerSource,
@@ -26,6 +25,7 @@ from instro.unstable.awg.types import (
     Triangle,
     Waveform,
 )
+from instro.lib.transports import VisaConfig
 
 _ARB_SAMPLES = (0.0, 0.5, 1.0, -1.0, 0.25, -0.25, 0.75, -0.75, 0.125)
 
@@ -34,7 +34,7 @@ _NO_ERROR = '0,"No error"'
 
 @pytest.fixture
 def keysight_visa_cls() -> Iterator[MagicMock]:
-    with patch("instro.unstable.awg.drivers.keysight_33521b.VisaDriver", autospec=True) as cls:
+    with patch("instro.awg.drivers.keysight_33521b.VisaDriver", autospec=True) as cls:
         yield cls
